about re-layout cpmt
