+++ README.md
@@ -22,10 +22,10 @@ src/
     ├── types/    # Типы и интерфейсы
     ├── ui/       # UI компоненты (кнопки, инпуты и т.д.)
     │   ├── components/
-    │   ├── layout/ # Шаблоны страниц 
+    │   ├── layout/ # Шаблоны страниц
     │   └── sections/ # Компоненты секций страниц
     │       ├── catalog/
-    │       └── ...      
+    │       └── ...
     └── utils/    # Утилиты и вспомогательные функции
 ```
