+++ src/shared/ui/sections/about/index.ts
@@ -1,0 +1,5 @@
+export {default as Hero} from "./Hero"
+export {default as Mission} from "./Mission"
+export {default as Team} from "./Team"
+export {default as History} from "./History"
+export {default as Stats} from "./Stats"
